Refactor out BaseBoundarySystem
