+++ Assets/MixedRealityToolkit/Providers/BaseBoundarySystem.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c0d5e9b1162c15046a015ddbb0901ea2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
